+++ jinahub/crafters/SpellChecker/.gitignore
@@ -1,0 +1,1 @@
+model.pickle
